requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
nltk>=3.6.0
Sastrawi>=1.0.1
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.3.0
streamlit>=1.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.2.0
